refactor: improve sales system (#33)
* refactor: cleaning alembic versions

* refactor: change order date column type

* feat: reset alembic to initial schema

* refactor: add email validation when creating admins

* feat: add delete all method for users

* refactor: clear db before populating

* fix: add shopping cart service as one of the order service parameters

* fix: OrderServiceDep requires shopping cart service dep

* feat: alembic migration accept null as order date

* feat: using Numeric for monetary values

* feat: add state column for orders

* fix: rectify alembic migration for state column enum

* feat: review table

* feat: calculating star mean

* feat: endpoint to get reviews from product

to do: -rest of endpoints (post, delete, get by buyer, endpoint that returns true if buyer bought and received something?)
-aditional logic to ensure that only people that have bought the product can comment

* feat: updated  review so that buyers can only add a review if they have purchased the product

* feat: endpoints reviews

* feat: alembic stuff

* feat: paralelización de populate_db

* feat: add endpoints for orders

---------

+++ Back/app/models/orders/product_line.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Float, ForeignKey, Integer
+from sqlalchemy import ForeignKey, Integer, Numeric
 from sqlalchemy.orm import relationship, mapped_column
 
 from app.base import Base
@@ -21,7 +21,7 @@ class ProductLine(Base):
         ),
     )
     quantity = mapped_column(Integer, nullable=False)
-    subtotal = mapped_column(Float, nullable=False)
+    subtotal = mapped_column(Numeric(10, 2), nullable=False)
 
     order = relationship("Order", back_populates="product_lines")
     seller_product = relationship("SellerProduct", back_populates="product_lines")
